task_manager/requirements.txt.txt
Django==3.2.12
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2